File Structure
1) Changed File Structure.

2) worker/analysis.py uses now external_data.get_current_table

3) .gitignore added
