+++ worker/.gitignore
@@ -1,0 +1,4 @@
+*.pem
+.env
+__pycache__/
+.python-version
